+++ tests/testthat.R
@@ -1,4 +1,5 @@
 library(testthat)
 library(CBTF)
 
+daemons(2)
 test_check("CBTF")